+++ catsmat/IMUSANT_Tests/Analysis/IMUSANT_processing_Tests.cpp
@@ -467,15 +467,15 @@ TEST_F(IMUSANT_processing_Tests, FindMelodicSegments_LBDM_Test1)
     ASSERT_EQ("Bass", t2_bass->getPartName());
     ASSERT_EQ("Piano", t3_piano->getPartName());
     
-    vector<float> t1_soprano_strengths = segmented_parts[0]->getOverallLocalBoundaryStrengthProfile();
-    vector<float> t1_alto_strengths = segmented_parts[1]->getOverallLocalBoundaryStrengthProfile();
-    vector<float> t1_tenor_strengths = segmented_parts[2]->getOverallLocalBoundaryStrengthProfile();
-    vector<float> t1_bass_strengths = segmented_parts[3]->getOverallLocalBoundaryStrengthProfile();
-    vector<float> t2_soprano_strengths = segmented_parts[4]->getOverallLocalBoundaryStrengthProfile();
-    vector<float> t2_alto_strengths = segmented_parts[5]->getOverallLocalBoundaryStrengthProfile();
-    vector<float> t2_tenor_strengths = segmented_parts[6]->getOverallLocalBoundaryStrengthProfile();
-    vector<float> t2_bass_strengths = segmented_parts[7]->getOverallLocalBoundaryStrengthProfile();
-    vector<float> t3_piano_strengths = segmented_parts[8]->getOverallLocalBoundaryStrengthProfile();
+    vector<double> t1_soprano_strengths = segmented_parts[0]->getOverallLocalBoundaryStrengthProfile();
+    vector<double> t1_alto_strengths = segmented_parts[1]->getOverallLocalBoundaryStrengthProfile();
+    vector<double> t1_tenor_strengths = segmented_parts[2]->getOverallLocalBoundaryStrengthProfile();
+    vector<double> t1_bass_strengths = segmented_parts[3]->getOverallLocalBoundaryStrengthProfile();
+    vector<double> t2_soprano_strengths = segmented_parts[4]->getOverallLocalBoundaryStrengthProfile();
+    vector<double> t2_alto_strengths = segmented_parts[5]->getOverallLocalBoundaryStrengthProfile();
+    vector<double> t2_tenor_strengths = segmented_parts[6]->getOverallLocalBoundaryStrengthProfile();
+    vector<double> t2_bass_strengths = segmented_parts[7]->getOverallLocalBoundaryStrengthProfile();
+    vector<double> t3_piano_strengths = segmented_parts[8]->getOverallLocalBoundaryStrengthProfile();
     
     ASSERT_EQ(17, t1_soprano_strengths.size());
     ASSERT_EQ(12, t1_alto_strengths.size());
